restructure the gql sources
